fix/STAYBNB-130 auth and Navbar triggering useSearchParams errors on app build
