+++ src/app/(site)/layout.tsx
@@ -1,4 +1,4 @@
-import Navbar from "@/components/Navbar";
+import NavbarWrapper from "@/components/NavbarWrapper";
 
 export default async function SiteLayout({
   children,
@@ -8,7 +8,7 @@ export default async function SiteLayout({
   return (
     <div className="flex flex-col h-screen">
       <header>
-        <Navbar />
+        <NavbarWrapper />
       </header>
       <main className="flex-1 flex-grow overflow-auto flex flex-col items-center w-full h-full">{children}</main>
     </div>
